create useAuth custom hook
